Update number.py
When trying to change the refrigerator or freezer temperature, it reported an error and did not work.

Making this change to the code solved the problem, and now increasing or decreasing the temperature works correctly.

Could you please correct it in the original file?

+++ custom_components/smartthingswasher/number.py
@@ -103,7 +103,7 @@ CAPABILITY_TO_NUMBERS: dict[
                 use_temperature_unit=True,
                 command=Command.SET_COOLING_SETPOINT,
                 range_attribute=Attribute.COOLING_SETPOINT_RANGE,
-                int_type=STType.FLOAT,
+                int_type=STType.INTEGER,
             )
         ]
     },
